challenge_demo/requirements.txt
alabaster==0.7.12
appnope==0.1.0
Babel==2.7.0
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.1
Django==2.2
django-cors-headers==3.1.1
django-extensions==2.2.5
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
future==0.18.2
geocoder==1.38.1
idna==2.8
imagesize==1.1.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.2
pytz==2019.3
ratelim==0.1.6
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
